requirements.txt
asgiref==3.10.0
black==25.9.0
click==8.3.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pytokens==0.3.0
sqlparse==0.5.3
